Add option to collect coverage artifacts
When a repo's `bin/test` includes code coverage metrics, we don't really
get much assistance in tracking down cases where the build has been
rejected due to a drop in coverage below our desired minimum.

I encountered this on a repo where coverage was 99.8%, triggering a
failure because the project minimum was 100%.

This commit adds a new optional `store_artifacts` step, which is run if
the orb's `collect-artifacts` flag is set to true (default is false for
backwards compatibility).

If this flag is switched on, the Simplecov HTML output that is output to
`<project>/coverage/` should become available to view from a Circle CI
run's `artifacts` tab.